Refactor: genericise GameObject, Building update(); Building, Particle to modules
