+++ src/game_object.py
@@ -75,12 +75,13 @@ class GameObject(pg.sprite.Sprite):
                 self.rect.y += self.speed * dy / dist
             self.rect.clamp_ip(pg.Rect(0, 0, MAP_WIDTH, MAP_HEIGHT))
 
-    def update(self) -> None:
+    def update(self, *args, **kwargs) -> None:
+        super().update(*args, **kwargs)
         self.move_toward()
         if self.cooldown_timer > 0:
             self.cooldown_timer -= 1
 
-    def draw_health_bar(self, screen: pg.Surface, camera: Camera) -> None:
+    def draw_health_bar(self, surface_: pg.Surface, camera: Camera) -> None:
         health_ratio = self.health / self.max_health
         if not self.under_attack and health_ratio == 1.0:
             return
@@ -88,13 +89,13 @@ class GameObject(pg.sprite.Sprite):
         bar_width = max(10, int(self.rect.width * health_ratio))
         screen_rect = camera.apply(self.rect)
         pg.draw.rect(
-            screen,
+            surface_,
             (0, 0, 0),
             (screen_rect.x - 1, screen_rect.y - 16, self.rect.width + 2, 10),
         )  # Background
-        pg.draw.rect(screen, color, (screen_rect.x, screen_rect.y - 15, bar_width, 8))
+        pg.draw.rect(surface_, color, (screen_rect.x, screen_rect.y - 15, bar_width, 8))
         pg.draw.rect(
-            screen,
+            surface_,
             (255, 255, 255),
             (screen_rect.x, screen_rect.y - 15, self.rect.width, 8),
             1,
